Refactor package structure.
