+++ elasticjob-infra/elasticjob-restful/src/main/java/org/apache/shardingsphere/elasticjob/restful/pipeline/ExceptionHandling.java
@@ -28,8 +28,8 @@ import io.netty.handler.codec.http.HttpResponseStatus;
 import io.netty.handler.codec.http.HttpUtil;
 import io.netty.handler.codec.http.HttpVersion;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.shardingsphere.elasticjob.restful.ExceptionHandleResult;
-import org.apache.shardingsphere.elasticjob.restful.ExceptionHandler;
+import org.apache.shardingsphere.elasticjob.restful.handler.ExceptionHandleResult;
+import org.apache.shardingsphere.elasticjob.restful.handler.ExceptionHandler;
 import org.apache.shardingsphere.elasticjob.restful.Http;
 import org.apache.shardingsphere.elasticjob.restful.serializer.ResponseBodySerializer;
 import org.apache.shardingsphere.elasticjob.restful.serializer.ResponseBodySerializerFactory;
