+++ elasticjob-infra/elasticjob-restful/src/main/java/org/apache/shardingsphere/elasticjob/restful/pipeline/HttpRequestDispatcher.java
@@ -23,11 +23,11 @@ import io.netty.channel.ChannelInboundHandlerAdapter;
 import io.netty.handler.codec.http.FullHttpRequest;
 import io.netty.handler.codec.http.HttpMethod;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.shardingsphere.elasticjob.restful.HandleContext;
-import org.apache.shardingsphere.elasticjob.restful.Handler;
-import org.apache.shardingsphere.elasticjob.restful.HandlerMappingRegistry;
-import org.apache.shardingsphere.elasticjob.restful.HandlerNotFoundException;
-import org.apache.shardingsphere.elasticjob.restful.MappingContext;
+import org.apache.shardingsphere.elasticjob.restful.handler.HandleContext;
+import org.apache.shardingsphere.elasticjob.restful.handler.Handler;
+import org.apache.shardingsphere.elasticjob.restful.handler.HandlerMappingRegistry;
+import org.apache.shardingsphere.elasticjob.restful.handler.HandlerNotFoundException;
+import org.apache.shardingsphere.elasticjob.restful.mapping.MappingContext;
 import org.apache.shardingsphere.elasticjob.restful.RestfulController;
 import org.apache.shardingsphere.elasticjob.restful.annotation.ContextPath;
 import org.apache.shardingsphere.elasticjob.restful.annotation.Mapping;
