+++ elasticjob-infra/elasticjob-restful/src/test/java/org/apache/shardingsphere/elasticjob/restful/RegexPathMatcherTest.java
@@ -17,6 +17,8 @@
 
 package org.apache.shardingsphere.elasticjob.restful;
 
+import org.apache.shardingsphere.elasticjob.restful.mapping.PathMatcher;
+import org.apache.shardingsphere.elasticjob.restful.mapping.RegexPathMatcher;
 import org.junit.Test;
 
 import java.util.Map;
