+++ elasticjob-infra/elasticjob-restful/src/test/java/org/apache/shardingsphere/elasticjob/restful/RegexUrlPatternMapTest.java
@@ -17,6 +17,8 @@
 
 package org.apache.shardingsphere.elasticjob.restful;
 
+import org.apache.shardingsphere.elasticjob.restful.mapping.MappingContext;
+import org.apache.shardingsphere.elasticjob.restful.mapping.RegexUrlPatternMap;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
